Moved menu styles to different file to maintain clean code base
