+++ client/src/components/Navbar/Menus/Unauthenticated/index.tsx
@@ -1,24 +1,15 @@
 import React from 'react';
-import { Grid, makeStyles } from '@material-ui/core';
+import { Grid } from '@material-ui/core';
 import { useHistory } from 'react-router';
 import { Link } from 'react-router-dom';
-
-const useStyles = makeStyles({
-  button: {
-    color: 'white',
-    fontWeight: 'bolder',
-    border: '1px solid white',
-    padding: '15px 40px',
-    borderRadius: 0,
-    letterSpacing: 1,
-    textDecoration: 'none',
-  },
-});
+import useStyles from './useStyles';
 
 export const index = (): JSX.Element => {
   const classes = useStyles();
   const history = useHistory();
-  const [location, setLocation] = React.useState(history.location.pathname);
+  const [location, setLocation] = React.useState<string>(history.location.pathname);
+
+  console.log('history:', history.location.pathname);
 
   return (
     <Grid item>
