+++ client/src/components/Navbar/Menus/Unauthenticated/useStyles.tsx
@@ -1,0 +1,15 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles({
+  button: {
+    color: 'white',
+    fontWeight: 'bolder',
+    border: '1px solid white',
+    padding: '15px 40px',
+    borderRadius: 0,
+    letterSpacing: 1,
+    textDecoration: 'none',
+  },
+});
+
+export default useStyles;
